requirements.txt
numpy
pandas
scikit-learn
tensorflow
flask
psycopg2-binary
python-dotenv
pyyaml
joblib
